backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
aiosqlite==0.21.0
greenlet>=3.0.0
# asyncpg==0.29.0  # Production only (PostgreSQL). Install separately when deploying.
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.2
thefuzz==0.22.1
scikit-learn==1.5.2
python-multipart==0.0.9
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
fake-useragent==1.4.0
tqdm==4.66.1
python-dotenv==1.0.1
